Eliminate "Endlike Headers"
An early concept of how to make it possible to compact small array
payloads directly into the node used the area for tracking an
allocation instead for a value cell.  For this to work, the bits after
the cell had to be able to act as an "END" signal.

But the system is moving to where there is no need for end signals, and
arrays just enumerate until they hit the tail as marked by a pointer.
With this change, endlike headers are no longer needed.