Make array termination a debugging feature
This switches away from using END to terminate arrays, and instead
only allocates the extra cell of space for the terminator in debug
builds when requested...writing a "trash" cell instad.

The END state is reserved for other purposes, such as the end of a
variadic feed or an empty state of a cell to use for "invisible"
detection in evaluation.