Commit to END marker usage for feed termination
With END markers no longer in arrays, the question came up if it was
worth it to dereference the feed pointer to notice if it was an END
or not.  If the feed was just nullptr, it could be tested faster.
Also, `option(REBVAL*)` could help signal that those dealing with
f_value needed to test it for END.

However, trying this out settled that it isn't worth it.  So END is
the better idea.

This also gives sys-end.h its own file, and moves the sys-nulled.h
and sys-track.h out of the datatypes subfolder of include.