Make panic() more informative in the web build
The panic() feature was only presenting debug file and line numbers in
the debug build.  Making the web build a full debug build is cost
prohibitive due to the bloat of all the asserts.

So this makes panic()'s file and line numbers (as well as other info)
controlled by DEBUG_FANCY_PANIC.  It can be used in performance
oriented builds (such as callgrind) as well as the debug build.