Fix LET in bootstrap shim
Because the bootstrap executable does not have the ability to create
bindings dynamically at runtime with LET, it is faked by making the
FUNC creation routine search for LETs and add them as `<local>`s.

However, those LETs are left in the body and need a runtime behavior.
This adjusts that behavior to erase WORD! values, vs. just being
invisible (which works for SET-WORD!, but not WORD!)