Address reuse of LET patches in virtual binding chains
The dynamic LET concept of making new variable bindings at runtime
uses a trick where a singular array node acts as a context -and- a
virtual binding patch.  The MISC() and INODE() fields are used for the
same purposes as other patches...but the available LINK() node holds
the variable's symbol.

This idea was hacked on as a test, but didn't have exhaustive handling
for when virtual bind chains had to be composed together.  It isn't
an especially common situation, but can be forced with:

    block1: do [let x: 10, [x + y]]
    block2: do compose/deep [let y: 20, [(block1)]]
    30 = do first block2

A more convoluted version of this situation arose in the watchlist, and
there was no handling for it.

This commit adds basic handling--though the approach continues to be
an exploratory one.