Fix virtual bind patch merge case for lambdas
A problem came up in the console when a GROUP! was started but not
closed, due to nesting lambda functions.  This led to finding that
a case which was thought to be uncommon enough to assert on.

The problem actually arose not from a very strange case, but just from
a redundant specification that occurs when virtual binds are used in
lambdas on top of a block and then run the block that the binding was
derived from.  It probably shouldn't do that, but this addresses the
issue by making such derivations work anyway.