Fix REEVAL frame reuse case, split out tests
There was a situation in REEVAL where it was reusing the frame, and
the frame state still said it was reevaluating...causing that reuse
to make an infinite loop.

This addresses that, as well as breaking out the REEVAL tests from the
DO test file.  (Since REEVAL of an ACTION! takes place of the variadic
behavior for DO on an action, the relevant tests for DO were put there,
so subsequent REEVAL tests wound up there as well.)

Also cleans up an unused string from the panic routine.