Avoid double-specification of LET in REEVAL
When performing a REEVALUATE, the value has been derelativized already.
It may have been fetched from a variable or evaluation, in which case
the LET bindings being added by the stream of evaluation should not
be applied.  Had it been literal, then it would already be applied.

This change looks at the state the frame is in before applying the
specifier.  If it in a reevaluating state, it assumes the value has
already been specified.