Make Multiple SET in PARSE an error
This restores the Rebol2 behavior of making a SET with a rule that
resolves to more than one item an error.

It also restores a behavior of making a SET with a rule that resolves
to nothing (e.g. an OPT rule that had no match) to give NULL.  (This
is better than historical NONE, because you couldn't tell if that was
actually matching a NONE! value in a block parse or not.)

Offhand, I recalled one case where the behavior of leaving a variable
"as is" in the opt case had come up.  This adjusts it to take advantage
of the null set.