Workaround binding problem in multiple returns
The multiple return code is sketched out by way of some slow libRebol
code (that needs to be replaced).

However it demonstrated a problem with using literal block values in
binding.  Without the protection afforded by getting a variable through
a word lookup, using a block as data can be affected by the binding
of an enclosing function.

This works around the issue by using LET to create a local variable.
That's an undesirable solution...but this falls under the general
set of issues related to "binding needs better answers".