Initial implementation of UPARSE (usermode PARSE)
This is a draft implementation of a new architecture for PARSE, based
on a style more similar to parser combinators.

It's very inefficient, but is already demonstrating a fairly good
amount of coverage of parse behaviors...with a more robust idea of
how to put together rules without needing to put them in blocks.  For
instance, it can do things like:

    parse ["aaaa"] [into some 2 "a"]