Simplistic combinator-based COLLECT+KEEP in UPARSE
The combinator-based approach to parsing breaks each parse keyword or
datatype behavior out into a "parser function".

This is a simple implementation of COLLECT and KEEP combinators that
collaborate with the BLOCK! combinator to do rollback.  It would be
preferable to have some kind of "generic rollback facility", but this
is more or less what the C code was doing to implement the feature.

The PARSE as a whole maintains a block that it collects, and marks the
high-water point on each block recursion...so that if an alternate
fails it just drops the mark down.