Switch PARSE's `collect x` pattern to `x: collect`
The initial design of COLLECT followed PARSE's pattern of naming a
variable after the keyword as the place to collect into.

A more forward-looking pattern is to use SET-WORD!s as they would seem
to be intended, by putting them on the left of the combinators that
are generating the value for them to store.  This is adopted by the
new UPARSE as the default (though it can be overridden).

This applies that rule to the native PARSE code for its implementation
of COLLECT with rollback.