Make command line arguments relative to launch path
This makes it so that if you use relative paths in rebmake.r on the
command line, those paths are relative to the directory you were in
at the shell prompt when you ran the interpreter.  Previously it was
relative to the location of %make.r

https://forum.rebol.info/t/1541

Also: in order to gloss over the differences between %/C/.. on Windows
and no drive-letter on absolute paths in Linux, source files are named
in command lines through relative paths.  This meant repo-dir became
relativized.  That change is now done after command-line processing.

Switches section labels to strings, as that is the likely best
behavior for the section-dividing `===`.