Try adding handling for -D macros with double quotes
This is an edge case triggered by mbedTLS due to how it lets you
add config files.

Make the C99 command line processing put back in backslashes to escape
double quotes.

Also, add this as part of the fib.r test, just to bring the issue to
the forefront since it's easy enough to do.