Deal with interim case-sensitive issue in %prep-main.reb
Picking out of a BLOCK! by WORD! was not failing when the word wasn't
present, which was different from how object picking worked.  This
glossed over a bug in %prep-main.reb

Too much code seems to depend on the case insensitive picking of words
out of blocks, and nulling if not there.  Punt on it for now, but fix
the particular bug and add some tests to capture what the behavior is.