Make the nmake build option pass /DEBUG on link line
GCC and Clang don't need you to pass -g to the linker if the obj files
themselves contain debug information.  It seems that MSC's linker does
need you to pass it.

(Trying to figure Rebmake out enough to switch depending on Visual
Studio to being able to use VSCode + NMAKE as a first step of starting
Rebmake over from scratch.)