Don't capture matrix variables into environment vars
Originally the matrix variables were captured into environment
variables to try and make the GitHub Actions scripts more portable.

However, it is turning out that the scripts are pretty brief and
understandable (compared to what was on Travis).  So this just uses
the matrix variables in the GitHub syntax, as it would be easy
enough to see how to change in other systems.