Add Android Workflow
This builds a "lowest common denominator" NDK version (r13) as
well as a recent NDK (r21).

It does the older version on Ubuntu.  The version is old enough
that it uses a long-deprecated gcc toolchain.

The newer version is built on Mac...to demonstrate that you can
cross-compile from multiple platforms and have Rebmake detect
which to use.  But also because the Android headless emulator
on GitHub actions is only working on Mac VMs at the moment.