Android NDK Version Detection in Configs
This takes advantage of the new feature of config inheritance to make
the detection of the NDK directory and version common between the
Android native and Android cross-compile builds.

It detects whether to use gcc or clang for the compiler, and also
uses unified headers for the sysroot of compilation for newer versions
of the NDK.

It's a significant amount of work, and the comments attempt to explain
all the details.  Being able to still build using the old and deprecated
API Levels and tools is done as a distinguishing feature...so that we
could still develop for older phones as well as the newest ones using
the tool.