Add `Inherits:` Header Option for Build Configs
Sometimes there are configuration options that are common to multiple
builds, beyond just inheriting the `default-config.r`.

There was a non-functional version of inheritance through having an
object field for a `config`.  Besides being in-band of the actual
config data, it did not run the innermost configs first...meaning that
the outermost configs could not build on properties discovered by the
inner ones.

This changes the approach to go through a discovery phase to build a
stack of configs.  This way it can run the innermost ones first and the
outermost ones last.  It also makes the Inherits: property part of the
file header, so it doesn't show up in the data itself.