Change NDK variable to ANDROID_NDK_ROOT
There does not appear to be a standard for the name of an environment
variable where the NDK is installed.  Some code on the internet seems
to use $NDK, while the original Ren-C Android build scripts assumed the
name was $ANDROID_NDK

With multiple environment variables in play, it's not clear what this
represents...it could be the version, or whether to install it or not.
So it had been changed to ANDROID_NDK_DIR to be more communicative.

However, when trying out the emulator in the android SDK (Java+tools)
it turns out there is a standard for that: ANDROID_SDK_ROOT

This would make the best choice appear to be ANDROID_NDK_ROOT as it
communicates that it is a directory and lines up with what the
binaries in the SDK look for in its variable.