Android Emulator To Online Test ARM Build Product
x86-based Android-11 devices are able to translate ARM
binaries on the fly.  This not only means that hardware devices
based on Intel Atom can use ARM binaries, but it also means
that x86 emulators can be used to run them.  This makes it more
practical to speed up the emulators when they are running on
x86_64 hosts (such as the MacOS container on GitHub Actions)

This commit adds an Android-11 emulator to the build process.
So an executable which is cross-compiled can actually be tested
as part of continuous integration.

See the detailed comments for more information.