Remark on the standardization of bash on all platforms
Bash is geared up to use the -e setting, which I actually did not know
about.  This affects how commands are written, in terms of not breaking
them into steps solely for the purpose of clearer errors.

Documents the behavior as it's not something obvious to those who are
reading the scripts and familiar with bash's "keep going" default.