Add --fragment run mode motivated by GitHub Actions
This makes an execution model for code from a file that doesn't change
the directory...and wouldn't expect a header.  This permissive mode
also loosens up to allow CR+LF sequences on Windows.

It was added in order to work better with GitHub Actions, which allows
running a custom shell:

    step:
      shell: r3 --fragment {0}
      run: |
        print "Hello, Ren-C As Shell World!"
        print ['no-apostrophe-backslashes "...no quote backslashes"]

GitHub turns the code in the `run` portion of the YAML into a temporary
file, in a temporary directory, with CR LF characters on Windows.  We
get the wrong results and errors if we try using ordinary script
execution, so this addresses that.