Add Handling for COMMA! cases like [/a/, b., c]
If you tried to tack a comma onto the end of a BLANK!-tailed TUPLE! or
PATH! then it was trying to put the comma into the path itself, and
it's not one of the legal types for paths.

This detects the situation and properly adds the blank, ending the path
scan and letting the parent scan get the comma.