Android TCC Extension Revived, cloud ARM emulator test
This brings the TCC extension building up to date.  Changes to TCC
itself had made it not work with ARMv5 any longer, and the Android
NDK no longer supports GCC in current versions.

The process of cross-compiling TCC is isolated from the main build
scripts into a helper script that is part of the TCC extension.  This
script retains the ability to build using older GCC-based NDKs.  That
involved several hacks, and those hacks are now isolated to if you
are building using them.

(Comments in the files explain the laborious number of details that
had to be attended to in order to get this to work!)

Because the only Android emulator that works with any performance is
Mac-based, the cleanest way to test would be to build on a mac and
then test on that same container.  So this makes the build work on
Macs as well.

The emulated Android device does not have a GNU toolchain installed,
so it lacks the libc or headers necessary for functions like printf().
This takes advantage of the new feature of COMPILE of /NOSTDLIB to
make sure that a compilation test can run in the emulator.