Rethink CHANGE to work with value-bearing rules
The CHANGE mutating primitive has a first rule that fits into the
PARSE model because it is a rule that matches the extent of what it is
to replace.  However, the second argument is a bit more puzzling...as
it is literal data and not a rule.

This tries to cohere the model by making the second argument have to
be a "value-bearing rule".