Add RETURN keyword to UPARSE
This makes it a bit easier to get results out of UPARSE, by reviving
the RETURN keyword from R3-Alpha.

The RETURN keyword was removed due to some discomfort about the
inconsistency it brought to the interface.  Many of those issues have
been mitigated somewhat, and there are now more good reasons (e.g.
COLLECT and GATHER) to want to make the PARSE return an aggregate
structure more easily.