Add UTF-16 support for Windows
Initially the code just supported UCS-2, using UTF-8 on Windows.  Now
that Emoji are pervasive, people might use them in places like the
Windows command line (in fact, it was used in the HTTPD demo to pass a
high codepoint to use in the server string, meaning that demo would
not work on Windows).

This adds the fairly easy support for surrogate pairs.