Try integrating action use to install executable
Running as a shell requires putting the binary in the path.  MacOS
isn't allowing even sudo to put things in /usr/bin, but rather than
figure out how to work around that this goes ahead and gets some
benefits from the split of the deploy and greenlight step by testing
the GitHub action for installing Ren-C and running scripts.