Restore WORD! spellings vs. getting from binding
As an attempt to clear space in WORD! cells for use in caching, the
spelling was assumed to be derivable from the binding.  This was a
questionable idea that caused several complexities...like making
case-insensitive behavior more difficult.

The strategy was not thought to be ideal or final, but reclaiming the
bits in the word would be needed by any superior caching strategy so
it was tried.

However, this got in the way of a new idea of making words bindable
to contexts without having a specific variable backing them.  That
idea is conceived to promote two important concepts: the freer use
of words to represent text, e.g.:

    === THIS IS A SECTION HEADING ===
    echo [This creates a lot of @variables, otherwise!]

But also this would allow global variables to hold references without
actually declaring those values.  This would be able to notice stray
writes, and have special constructs for forcing a declaration to
be created if that was what was intended.

These ideas are incompatible with having the word coming from the
binding.  So this removes the use of binding to get the spelling...
which loses the VAL_WORD_CACHE() slot.

However, the portion of the cache that was "some of the index bits"
remains.  This is necessary if the bindings of QUOTED! are going to
be stored uniquely for multiple instances of the same quoted cell.
While this isn't much, it's at least something that can be used in
a rethought virtual binding cache scheme.