Eliminate Reb_Series_Base
The C++ build uses inheritance for the purposes of type checking, e.g.
to ensure that you don't pass a string where an array was wanted or
vice versa...while still letting them share an implementation type.
This means all the costs for the checks are paid at compile time, and
no difference in runtime from the C version.

In this set of checks, there was a notion that although contexts and
actions and maps are implemented on top of arrays, that is a detail
that should not be exposed widely in the system as it could change.
Hence you cannot pass a REBCTX to somewhere that a REBARR is needed,
you have to ask for the variable list explicitly through CTX_VARLIST()

While this is a good idea, it creates a bit of a hassle by not having
a common type for comparison purposes.  e.g. the binding of a word
can be to an action or a context or a virtual binding "patch".  But
if none of these are willing to say they are arrays, then you can't
test to see if `action == BINDING(v)` because you can't compare pointers
of completely unrelated types.

The solution that had been conceived for this was to have a Node type
(REBNOD) that was a base class for all header-bearing types.  However,
that included REBVAL and REBFRM... so if this is the only common
denominator between actions and arrays and contexts then you get
comparison ability but little else.  This makes the type undesirable
to use for something like the return result from BINDING().

After looking at the mechanical issues involved in trying to contort
REBCTX/REBACT/REBMAP into being "REBNOD but not REBARR", it looks to
be a better tradeoff to make them "REBSER but not REBARR".  This is
slightly more permissive by letting you use REBSER-based functions
like SER_USED() on a REBCTX...but there are very few relevant fully
generic SER-based functions any longer.  (Most things are subclassed
into REBSTR, REBSYM, REBARR...even getting the lengths.)

With REBSER as the common denominator, it makes more sense to return
REBSER from the BINDING(), and then in addition to doing comparisons
you also have access to LINK/MISC/FLAVOR checking.

This goes ahead and makes that change, which means it's no longer
necessary to go through some amount of contortion of multiple
inheritance to say that e.g. REBCTX was *not* a REBSER but *was* a
node...which had to do acrobatics in order to avoid losing the empty
base class optimization that kept it compatible with C.