Make UPARSE able to EMIT variables to callsite
This makes it so that if EMIT is performed when there is no GATHER in
effect, then UPARSE will actually declare those variables as if it
were LET statements for each variable.

    i: #i
    t: #t
    if true [
        uparse [1 <foo>] [emit i: integer!, emit t: tag!]
        assert [i = 1, t = <foo>]
    ]
    did all [
        i = #i
        t = #t
    ]

Also makes ADD-LET-BINDING more efficient by allowing you to pass in
the value of the variable that you are declaring.