Make TO and THRU value-bearing in UPARSE
The behavior of `set x thru "a"` was not very appealing in historical
Rebol.  Usually you needed to say `copy x thru "a"` to get the
desired result.

Initially Ren-C made TO and THRU non-value bearing, so they couldn't
be used with SET-WORD!, and `x: thru "a"` was just an error.  The
concept was that value-bearing things might be rare, in order to
be able to tell what parts of a block rule were intended to gather
data or not...hence the ACROSS (e.g. COPY) would call out spots that
were purposeful gathering of values.

But, it's probably more of a pain than help, so this goes ahead and
makes it so that the semantics of `x: to "a"` are the same as
`x: across to "a"`... so the value is the range of input across that
part of the series.