Make WHILE the default in UPARSE for ANY behavior
This is to test out the idea of replacing ANY with WHILE in UPARSE.

Historically WHILE was a slightly different version of ANY in PARSE,
which did not consider reaching the end to be grounds for terminating
the rule.  ANY would terminate, but considering the overall rule to
be successful...as if you had written `while [not end, ...]` instead
of `any [...]`.

Having two constructs for iteration that accepts 0 iterations as a
"succeeding rule" is somewhat confusing.  But also, ANY as it acts in
DO is very much like `rule1 | rule2 | rule3` in PARSE.  So iterating
on something is a fully different meaning...when that meaning seems
to be legitimate.

This tries out the idea of just using WHILE, to see if that is more
palatable.  ANY is kept as a synonym for now.

The distinction was subtle, and in the cases where it actually matters
it can be seen as more instructive to point out the NOT END in order
to possibly rethink the rule in a simpler way.  If it can't be
simplified, then pointing out the presence of the NOT END may be
clearer and better than