Add TYPESET! handling to UPARSE, more tests
There is still some amount of uncertainty surrounding the TYPESET!
data type.  But in order to try running a version of REWORD that uses
UPARSE in its implementation, it was needed.

This goes ahead and adds it, as well as that test.  Also includes a
test for breaking a string into escaped word portions (came from a
string interpolation test).