Update Encapping Code
The encap code had not been tested in some time, and as a result had
several undefined variables.

This brings it up to date with that.  It also goes ahead and simplifies
the interface to just work with ZIP files and the meaning of BLOCK! to
the ZIP dialect (which is still under definition).

Another repository is being built for tracking encapping specifically
as well as running tests.