Fix Redbol handling of functions
SET-WORD!s have been changed to deal with return results exclusively,
so TUPLE!s were changed to be locals.  (e.g. `.FOO` instead of `FOO:`)

The Redbol function translation relied on the SET-WORD! to deal with
the old style of refinements having arguments named differently from
the refinement.  This needed a couple of changes that had not been
done yet.