Build and test the rebol-server.apk for Android
This re-enacts @giuliolunati's demonstration of making an installable
.APK out of an Android native build of r3 with httpd.reb, bundled with
a Wasm build of the ReplPad.

The result is a package that can be installed and run offline.  Also
the native Ren-C can act as a backdoor into whatever functionality that
native Android could offer that the browser could not.  (A simple
file browser is included as an example.)

This commit takes the extra step of putting the built APK into the
Android emulator, and performing a test that causes the Wasm r3 client
to make a request to the Android-native r3 server.  The request leads
a file to be written that can be looked for by the host as proof that
the operations were both able to run and communicated with each other.