Add Redbol "no progress" rules for ANY and SOME
Ren-C is experimenting with the concept of changing the trio of ANY,
SOME, and WHILE--with the first two being stopped by no progress--
to just WHILE and SOME with neither having a progress requirement.

However, Redbol emulation needs to work with the old rules.  This
adds combinators to the UPARSE2 for that.