Change UPARSE WHILE and SOME to not synthesize result
In order to try and give a "sensible" result when you wrote things
like `uparse ... [x: while [...]]`, it would act as an implicit
COLLECT.  This is not how loops work in the language proper (except
in the case of MAP-EACH).  So this switches it around to just return
the last loop value-bearing rule result.

While this may not seem particularly useful, it's not the same as
being invisible...and may be at least as useful as the same property
of loops in DO.