Reform meaning of SYM-XXX!, allow result for GROUP!
Before the clear separation between "advanced input" and "synthesized
value", there was wariness about rules "seeing the result" of a
plain GROUP! evaluation.  If PARSE could "see the result" and it was
not an argument to an obvious value-taking routine like KEEP, then
what would "seeing" mean?  Running as a rule?

But with the separation and a new proposal that BLOCK!s actually can
bear values synthesized from their last result, it makes it seem
practical and normal that plain undecorated GROUP!s have their results
be "seen" by default, if they are not ELIDEd.

This makes the SYM-GROUP!/SYM-WORD!/etc. combinators seem more suited
to actually matching literally against an item in the input.  This
was the original idea that was pushed aside temporarily when the need
to inject values became more pressing.  But it became so pressing that
letting plain GROUP! do it made more sense.