Rethink UPARSE result concept
This changes UPARSE around to where the functions advance the input
through a separate variable (the remainder) that is not the main
return result of the combinator.  Instead, the main return result of
the combinator is used as "the result" of a parse rule expression.

Previously, returning a result was optional...and done only by request,
e.g. when a rule was being used with a SET-WORD!.  This created some
tension with the BLOCK! rule type, because it was less clear what the
result of such a rule should be.

By adopting a concept paralleling the "last result of a block evaluation
is the overall block's evaluation" (from ordinary DO-style evaluation),
this makes returning a result the norm...and the primary return value
of a function is that result.  This makes combinators more similar to
ordinary functions, perhaps leading to some kind of unification.