Lax equality does not consider QUOTED! equal to non
Historical Rebol allowed equality between LIT-WORD! and WORD! when
non-strict comparison was done:

    rebol2> (first [a]) = (first ['a])
    == true

This is questionable to begin with, but certainly questionable in Ren-C
where you don't want this to be true:

    (first [']) = null

If you want comparison to not consider quoting, you can do so by
dequoting the things you are comparing.  Having the core comparison
operator itself pretend the quote is "not there" is much more harmful
than it is helpful.