Remove QUOTED! test handling from MATCH
A very half-baked experiment of interpreting quote levels on tests
in MATCH was causing bugs in Redbol, because a QUOTED! was not
considered to be an ANY-VALUE! since its quote level did not match the
quote level of the typeset serving as a test.

The experiment was kept around because it didn't seem to do any harm,
but that's a pretty big harm.  So this gets rid of it.