Eliminate PARSE transcode test
With UTF-8 Everywhere, all strings are kept in UTF-8 format all the
time.  This makes it more feasible to invoke the scanner during PARSE
of a TEXT!.

A quick feature was added at a time when there was no clear protocol
on how a parse rule could have a "synthesized result".  So a special
return convention using R_IMMEDIATE was hacked in, that pushed the
value to be returned to the stack.

UPARSE is now paving the way for a much cleaner concept of how such
a thing would work.  So the broken demo is removed from PARSE.  (There
was only one demo in the tests:)

    parse "abc [d e f] {Hello}" ["abc" set b block! set s text!]
    b = [d e f]
    s = {Hello}