Change UPARSE to use "core return protocol"
This tweaks the return result of UPARSE so that it uses [] to mean
invisible, NULL to mean rule failure, and QUOTED! for all other return
values.

The result simplifies the code since the main result can be used to
cue success or failure based on "true null" while a successful null
is represented by a quoted null (').  It also means there's no need
to use GET-WORD!s to talk about the result.

As a result, ELIDE and COMMENT work.