Change SEEK to use rules instead of variables
Historical PARSE would use variable names quotedly in rules often
(e.g. the `X` in `COPY X [SOME RULE]`).

However, UPARSE is trying to avoid this idea (outside of legacy
emulation).  How hard a rule it will be remains to be seen, but the
only non-legacy combinator using quoting was SEEK.

When GROUP! rules are available, getting a seek of a position via
a variable is as easy as:

    seek (var)

This permits other possibilities blending in rules, like:

    seek [some "a", let pos: here, some "b", (pos)]

While no immediately obvious useful usages of a rule (besides group)
come to mind, they may exist...and having SEEK speak in the language
of rules instead of needing to quote its argument seems like a good
thing to try to see how it works out.