Change UNZIP to use UPARSE, Revisit Streaming Fix
This changes the UNZIP code to use UPARSE.  Although UPARSE is in the
prototype stage and very slow, what the ZIP code does with parsing is
relatively lightweight (processing header records for zip entries) so
it's good to get the experience and take advantage of the new features.

The most important feature exploited is to take advantage of the new
ability of rules to synthesize arbitrary values.  This means patterns
like:

    copy name-length: 2 skip
    (name-length: get-ishort name-length)

...can benefit from abstraction.  A rule can be created by the user
that not only matches 2 bytes in the input series, but synthesizes a
16 bit unsigned integer out of it:

    name-length: uint16-rule

Such abstractions were not previously possible within parse as rules,
but now it can be done simpy by having a BLOCK! rule which evaluates
to its last rule product.  In this case, a GROUP!...the group consumes
no input and always succeeds, returning its value:

    uint16-rule: [tmpbin: across 2 skip, (debin [LE + 2] tmpbin)]

In addition to taking advantage of this new feature, this commit also
goes through and does some tidying of changes that were made to
support a patch for working around zip files that were created with
streaming:

https://github.com/metaeducation/ren-c/issues/978

That code did more jumping around than it needed to, so this cuts the
jumps down to the minimum needed.  It also makes the coherence checks
more complete between the local file headers and the central directory.