USING and ADD-LET-BINDING bugfix
The idea of being able to import an object's fields as if they had
been defined with LET is an intriguing one, first conceived to make
it easier to let captured values bubble up out of a PARSE without
needing to use locals-gathering.

    use uparse [1 <foo>] [
       gather [emit i: integer!, emit t: tag!]
    ]
    assert [i = 1, t = <foo>]

This defines the construct as USING in order to not conflict with
historical USE or now.  But it looks like if it pans out, it would be
a good fit for taking the name.