New return conventions: UPARSE vs. UPARSE? vs UPARSE*
This reimagines the return results of parsing to where whatever the
rule block synthesizes is the default result.

Simply wanting to know if it logically matched to end or not is done
by means of UPARSE?, which avoids issues of if the rules matched but
evaluated to null.  Being conditioned to know the difference is
something users would have to be aware of.

The lower-level UPARSE* returns the result but does not enforce
reachig the end.  This makes it possible to get at the partial parse
progress location by means of ending the rule with HERE.