FURTHER Parser Combinator for Enforcing Advancement
R3-Alpha's parse tried to codify advancement of input by parse rules
in order to make infinite loops turn into failures more often.  This
caused many problems, so the advancing rule was limited to just ANY
and SOME...with a non-advancing form of ANY called WHILE.

But that was confusing.  UPARSE removed the advancement rule entirely,
but this adds a FURTHER combinator that can be used with any rule
to demand that it make progress on the input, or its success will
be turned into failure.  This can be used in isolation or with a
loop construct (`while further [...]`)