Make END return series, instead of being invisible
The idea behind why END was thought to be useful as being invisible
(instead of the end series position) isn't necessarily a great
idea.  It complicates some rules that are attempting to evaluate to
values, e.g. TO END expects you to do any ELIDE-ing on the TO...and
can't return the invisible END value.

These issues need more thought; for the moment, do the more obvious
thing and make END evaluate to the end series position (the way that
HERE evaluates to the current input position).