Eliminate ISOTOPE? Variable Testing
The idea behind the @LITERAL parameter convention and other related
changes is to make it so that the NULL-2 isotope is fully transient;
and is never reflected in variables.

But the console printing of return values was still based on the idea
of reading the isotope bit out of a variable via ISOTOPE?.  That
routine needed to be removed.

Once it was removed, the console had to be updated.  In the process,
it got the ability to discern between invisibility (hitting enter with
no evaluative product) and NULL.