Fix Make_Invokable for invisibles, WORD!s
The concept behind making an invokable from a feed is to return either
an action or a QUOTED! value that will produce the desired effect of
running the ensuing code.

This adds support for invisible functions by producing a frame for the
"NIHIL" arity-0 invisible.  It also fixes a problem where if a WORD!
did not look up to a function, the value of the quoted word was being
returned instead of a quoted version of what the word looked up to.