Drop EITHER-MATCH for isotopes, NON does passthru
MATCH attempts to pass through values if they fit a type constraint.
This raises the possibility that the type constraint might be something
like LOGIC! with the input being false, thus one writes:

    value: false
    ...
    if match logic! value [
        print "This won't run, though you intended it to"
    ]

The non-truthy-or-falsey status of bad words was leveraged to return
the result ~falsey~ to stop these comparisons.

However, you could still use the THEN and ELSE tests to know if a
match happened:

     match integer! false then [print "This will run"]

And a lower-level operation called MATCH* could be used if you were
sure you knew the result.

    match* (expression) then x -> [
        print "You can be sure X matched...unless it's NULL"

But NULL presented a problem for this.  So EITHER-MATCH was created, to
give full coverage of matching nulls.

But in the current world, "null isotopes" can be used to resolve even
his issue.  An input of NULL that matches in the low level MATCH*
operation can return NULL-2 and trigger THEN.